deepseek-chat-app/whisper-service/requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai-whisper==20240930
torch==2.5.1
torchaudio==2.5.1
zhconv==1.4.3
